Rename the component and container to UserMessages
