+++ app/components/Navigation/index.js
@@ -23,6 +23,9 @@ function Navigation() {
         <li>
           <Link to="/contact">Contact</Link>
         </li>
+        <li>
+          <Link to="/messages">Messages</Link>
+        </li>
       </ul>
     </nav>
   );
