+++ app/containers/App/index.js
@@ -13,8 +13,8 @@ import { Switch, Route } from 'react-router-dom';
 import NotFoundPage from 'containers/Pages/NotFoundPage/Loadable';
 import AboutMePage from 'containers/Pages/AboutMePage/Loadable';
 import PortfolioPage from 'containers/Pages/PortfolioPage/Loadable';
-// import ContactMePage from 'containers/Pages/ContactMePage/Loadable';
-import ContactMePanel from 'containers/Pages/ContactMePanel/Loadable';
+import ContactMePage from 'containers/Pages/ContactMePage/Loadable';
+import ViewMessagesPage from 'containers/Pages/ViewMessagesPage/Loadable';
 
 import GlobalStyle from '../../global-styles';
 
@@ -24,7 +24,8 @@ export default function App() {
       <Switch>
         <Route exact path="/" component={AboutMePage} />
         <Route exact path="/portfolio" component={PortfolioPage} />
-        <Route exact path="/contact" component={ContactMePanel} />
+        <Route exact path="/contact" component={ContactMePage} />
+        <Route exact path="/messages" component={ViewMessagesPage} />
         <Route component={NotFoundPage} />
       </Switch>
       <GlobalStyle />
