+++ app/containers/Pages/ViewMessagesPage/reducer.js
